Use DATABASE_URL as default value of DSN
We don't want a user to hard-code a DSN in a project.yaml, because
that's insecure. DATABASE_URL is available to all actions that have
database (i.e.  network) access, but we don't want a user to pass `--dsn
$DATABASE_URL` either, because whether the string is evaluated depends
on job-runner's implementation. So we use DATABASE_URL as the default
value of DSN.

We add test_parse_args_with_defaults_from_environ rather than
parameterize test_parse_args because the number and length of the
parameters would have made the test hard to reason about.