Add `--dummy-data-file` argument to CLI
At present, an SQL Runner action breaks a local run (e.g. `opensafely
run run_all`, because it cannot connect to a database. This makes it
hard to test a pipeline that includes an SQL Runner action.

By adding the `--dummy-data-file` argument to the CLI, we make it easier
to test such a pipeline. Whilst a user must create a dummy data file
themselves, they can nevertheless check that any downstream actions,
which would otherwise expect an output from the SQL Runner action, work
as expected.

Closes #32