Ignore `--dummy-data-file` in production
This allows us to commit `--dummy-data-file` to *project.yaml*, such
that the associated SQL Runner action is run against dummy data when run
locally and when run in CI, but run against a database when run in
production.

You'll notice that whilst the functions called from *__main__.py* are
tested, the `if`/`else` within *__main__.py* is not. This follows the
advice within the Python documentation. For more information, see:
https://docs.python.org/3.10/library/__main__.html?highlight=__main__%20py#id1

I briefly considered making `--dummy-data-file` and `--dsn` mutually
exclusive with `ArgumentParser.add_mutually_exclusive_group`, but the
point here is to commit both to *project.yaml*.

Closes #46