Add `--log-file` and structured logs in JSON
structlog calls this "the simplest approach"; logging takes care of the
handlers and structlog takes care of the rendering. Whilst this means we
output structured logs in JSON format to both stdout and to a file
(conditionally), the alternative would be what structlog calls "the most
ambitious approach". It's not essential to text-format output to stdout
but JSON format output to a file, so we eschew ambition for simplicity.

For more information, see:
https://www.structlog.org/en/stable/standard-library.html#suggested-configurations