Raise RuntimeError in core of program
When I created #130, it bothered me that run_sql was doing more than it
said it was doing; it was running the SQL, but only when T1OOs were
handled correctly. Furthermore, I felt that although the type of error
was correct, the location wasn't: Surely a run-time error should be
raised in a run-time-like function? (i.e. in the core of the program)

To address these issues, I've moved where we raise RuntimeError into
main.main (a run-time-like function). I've also moved read_text out of
the if/else, so it is always called.