Add docker-build-for-os-cli recipe
This commit first modifies the docker-build recipe to build the
sqlrunner-dev image by default (and modifies the tag, build, and publish
workflow to ensure that the sqlrunner image is still built on successful
merge to main). It then adds a docker-build-for-os-cli recipe to tag the
sqlrunner-dev image, such that sqlrunner:dev is picked up by the
OpenSAFELY CLI (opensafely run, opensafely exec).

Why? So that we can make changes to a local SQL Runner, and test them
against a local study. Doing so is useful when you want to debug odd SQL
Runner behaviour in your local study.