Error if dsn and dummy-data-file are both missing
Because we often want to supply both of these arguments -- for local/CI
and production use -- we can't make them mutually exclusive. However, we
want to raise an error (or, at least, raise a more informative error) if
they are both missing, because it's impossible for SQL Runner to do
anything useful.