Rework `test_run_sql_include_statistics`
I wasn't clear what this was testing, so I reworked it. (I also renamed
it.) In the process, I also removed the "no IO statistics" branch.

I think this was testing that the log messages conformed to a schema:
that is, that the log messages had the expected keys. If that's the
case, then testing key-by-key will tell us which key is missing, if a
key is missing, rather than that a key is missing. An alternative
approach would be to refactor `parse_statistics_messages` to return a
dataclass, as then the expected keys would always be present. We can
leave doing so for another day, though.