Add more comprehensive end-to-end tests
SQL Runner has three command-line arguments that determine its
behaviour: `dsn`, `dummy-data-file`, and `output`. Each can take a
string or None. In this commit, we set `output` to a string (a path).
Consequently, SQL Runner's behaviour is as follows:

| dsn  | dummy-data-file | Behaviour                         |
|------|-----------------|-----------------------------------|
| dsn  | path            | Query DB, write results to output |
| dsn  | None            | Query DB, write results to output |
| None | path            | Write dummy data file to output   |
| None | None            | Write column headers to output    |

Previously, we tested the dsn/None case and the None/None case. However,
the end-to-end test for the None/None case didn't reflect the changes
that were introduced by #93. This commit adds an end-to-end test for
each case. However, the None/None test doesn't yet reflect the changes
that were introduced by #93.