Refactor `write_results`
This refactors the `write_results` function, and returns it to it's
signature prior to #160. We want to output to stdout, if an output file
isn't supplied. Whilst `sys.stdout` behaves like a file-like object, we
don't want to close it, as we would an output file. Rather than
accomplish this through type checking and try/finally, we instead use a
different context manager to output to `sys.stdout`. Hopefully doing so
makes the `write_results` function easier to reason about.