Update the version of mssql used in tests
Copying the commit message from the equivalent PR on ehrql:
https://github.com/opensafely-core/ehrql/pull/2120

The previous version was incompatible with the updated kernel now used
by Github Actions runners. See:
microsoft/mssql-docker#868

The fix has been backported as far as the 2019 image, but not to the
2017 image. Seeing that the 2017 image didn't match what was being run
in production in any case we might as well upgrade.

This requires a few small changes:

 * The `sqlcmd` tool has moved from `/opt/mssql-tools/bin` to
   `/opt/mssql-tools18/bin`. This is due to the newer image using the
   `mssql-tools18` apt package.

 * The `sqlcmd` invocation now requires the `-C` option to automatically
   trust the server certificate.